+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 <!--next-version-placeholder-->
 
+## v0.4.0 (2022-03-31)
+### Feature
+* --clear теперь флаг, добавлен флаг no-sequencer ([`0241df9`](https://github.com/devind-team/devind-django-core/commit/0241df92883eb5bc76e012e85ecc67830a679972))
+
+## v0.3.0 (2022-03-16)
+### Feature
+* Добавлен идентификатор удаленной группы в мутации ([`3f1c9d8`](https://github.com/devind-team/devind-django-core/commit/3f1c9d8bb1a4f48cd7436e16da83d5ab407bb163))
+
+## v0.2.1 (2022-03-09)
+### Fix
+* Fix validator in profile models ([`17445ed`](https://github.com/devind-team/devind-django-core/commit/17445edf013e9d33d3bb9d70eb56af94870f4f31))
+
+## v0.2.0 (2022-03-08)
+### Feature
+* Update devind_helpers ([`c9e836f`](https://github.com/devind-team/devind-django-core/commit/c9e836f932221def59afe99eebfc2151ee6fd1b1))
+
 ## v0.1.3 (2022-03-07)
 ### Fix
 * Add missing import ([`1c86ee5`](https://github.com/devind-team/devind-django-core/commit/1c86ee56589feaa2414ee94512389d52bd872d83))